Moves versions into specific entities as the pattern isn't any longer shared across all mappers.
